backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8

# PostgreSQL (use in production)
psycopg[binary]>=3.1
dj-database-url>=2.1

# S3 storage (optional; use for production or when AWS is configured)
django-storages>=1.14
boto3>=1.34
